Remove outdated munger.pl
The current version is in ~jefferis/bin.

The version here was no longer maintained.  I verified that it did
not contain any relevant changes compared to the version in
~jefferis/bin.  Note that I failed to locate exactly the same ancestor.
The closes match I found was munger.pl of commit
7f2a7ebe91b2f7f163ffef5f11b4358742271cd6 in ~jefferis/bin, which
seems to be functionally equivalent to the version that this commit
removes.